% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottery.R
\name{mylottery}
\alias{check my lottery}
\title{Will my lottery win or not?}
\usage{
mylottery(num_vec)
}
\description{
This function tells you whether you won the lottery or not.
}
\examples{

mylottery(c(2,14,28,33,19,43,23))
}
